requirements.txt
celery==5.4.0
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
psycopg2==2.9.9
python-dotenv==1.0.1
redis==5.0.4
whitenoise==6.6.0